+++ clients/admin-ui/src/features/common/PickerCard.tsx
@@ -5,31 +5,25 @@ import {
   Checkbox,
   CheckboxGroup,
   Flex,
-  Spacer,
   VStack,
 } from "@fidesui/react";
 import { ReactNode } from "react";
 
 const NUM_TO_SHOW = 5;
 
-const PickerCard = <T extends { id: string; name: string }>({
-  title,
+export const usePicker = <T extends { id: string; name: string }>({
   items,
   selected,
   onChange,
-  toggle,
 }: {
-  title: string;
   items: T[];
   selected: Array<string>;
   onChange: (newSelected: Array<string>) => void;
-  toggle?: ReactNode;
 }) => {
-  const itemsToShow = items.slice(0, NUM_TO_SHOW);
-  const numSelected = selected.length;
-  const allSelected = items.every((item) => selected.includes(item.id));
-
-  const showViewMore = items.length > NUM_TO_SHOW;
+  const allSelected =
+    items.every((item) => selected.includes(item.id)) && !!items.length;
+  const someSelected =
+    items.some((item) => selected.includes(item.id)) && !!items.length;
 
   const handleToggleSelection = (id: string) => {
     if (selected.includes(id)) {
@@ -47,6 +41,42 @@ const PickerCard = <T extends { id: string; name: string }>({
     }
   };
 
+  return {
+    allSelected,
+    someSelected,
+    handleToggleAll,
+    handleToggleSelection,
+  };
+};
+
+const PickerCard = <T extends { id: string; name: string }>({
+  title,
+  items,
+  selected,
+  indeterminate,
+  onChange,
+  toggle,
+  onViewMore,
+  numSelected,
+}: {
+  title: string;
+  items: T[];
+  selected: Array<string>;
+  indeterminate: Array<string>;
+  onChange: (newSelected: Array<string>) => void;
+  toggle?: ReactNode;
+  onViewMore?: () => void;
+  numSelected: number;
+}) => {
+  const itemsToShow = onViewMore ? items.slice(0, NUM_TO_SHOW) : items;
+
+  const { allSelected, someSelected, handleToggleAll, handleToggleSelection } =
+    usePicker({
+      items,
+      selected,
+      onChange,
+    });
+
   return (
     <Box
       p={4}
@@ -55,7 +85,6 @@ const PickerCard = <T extends { id: string; name: string }>({
       gap="4px"
       borderRadius="4px"
       boxShadow="0px 1px 2px 0px rgba(0, 0, 0, 0.06), 0px 1px 3px 0px rgba(0, 0, 0, 0.1)"
-      maxWidth="363px"
       fontSize="sm"
       data-testid={`picker-card-${title}`}
     >
@@ -71,6 +100,7 @@ const PickerCard = <T extends { id: string; name: string }>({
             onChange={handleToggleAll}
             colorScheme="complimentary"
             data-testid="select-all"
+            isIndeterminate={!allSelected && someSelected}
           >
             {title}
           </Checkbox>
@@ -89,25 +119,30 @@ const PickerCard = <T extends { id: string; name: string }>({
         ) : null}
         <VStack paddingLeft="6" fontSize="sm" alignItems="start" spacing="2">
           <CheckboxGroup colorScheme="complimentary">
-            {itemsToShow.map((location) => (
-              <Flex key={location.id} alignItems="center" gap="8px">
+            {itemsToShow.map((item) => (
+              <Flex key={item.id} alignItems="center" gap="8px">
                 <Checkbox
-                  key={location.id}
-                  isChecked={selected.includes(location.id)}
+                  key={item.id}
+                  isChecked={selected.includes(item.id)}
+                  isIndeterminate={indeterminate.includes(item.id)}
                   size="md"
                   fontWeight="500"
-                  onChange={() => handleToggleSelection(location.id)}
-                  data-testid={`${location.name}-checkbox`}
+                  onChange={() => handleToggleSelection(item.id)}
+                  data-testid={`${item.name}-checkbox`}
                 >
-                  {location.name}
+                  {item.name}
                 </Checkbox>
               </Flex>
             ))}
           </CheckboxGroup>
         </VStack>
-        <Spacer />
-        {showViewMore ? (
-          <Button size="xs" variant="ghost">
+        {onViewMore ? (
+          <Button
+            size="xs"
+            variant="ghost"
+            onClick={onViewMore}
+            data-testid="view-more-btn"
+          >
             View more
           </Button>
         ) : null}