+++ clients/admin-ui/src/features/locations/LocationPickerCard.tsx
@@ -1,26 +1,77 @@
-import { Flex, FormLabel, Switch } from "@fidesui/react";
+import { useDisclosure } from "@fidesui/react";
 import { useState } from "react";
 
 import PickerCard from "~/features/common/PickerCard";
-import QuestionTooltip from "~/features/common/QuestionTooltip";
-import { Location, Selection } from "~/types/api";
+import { Location, LocationGroup, Selection } from "~/types/api";
+
+import RegulatedToggle from "./RegulatedToggle";
+import SubgroupModal from "./SubgroupModal";
+import { getCheckedStateLocationGroup, isRegulated } from "./transformations";
+
+const SEARCH_FILTER = (data: { name?: string }, search: string) =>
+  data.name?.toLocaleLowerCase().includes(search.toLocaleLowerCase());
 
 const LocationPickerCard = ({
   title,
+  groups,
   locations,
-  selected,
+  selected: selectedLocations,
   onChange,
+  search,
 }: {
   title: string;
+  groups: LocationGroup[];
   locations: Location[];
   selected: Array<string>;
   onChange: (selections: Array<Selection>) => void;
+  search?: string;
 }) => {
+  const disclosure = useDisclosure();
   const [showRegulatedOnly, setShowRegulatedOnly] = useState(false);
+  const showGroupsAndLocations = !!search?.length;
+  const isGroupedView = groups.length > 0;
 
-  const filteredLocations = showRegulatedOnly
-    ? locations.filter((l) => l.regulation?.length)
-    : locations;
+  let locationsForView = locations;
+  if (groups.length) {
+    locationsForView = groups;
+  }
+  if (showGroupsAndLocations) {
+    locationsForView = [...locations, ...groups].sort((a, b) =>
+      a.name.localeCompare(b.name)
+    );
+  }
+  const regulatedLocations = showRegulatedOnly
+    ? locationsForView.filter((l) => isRegulated(l, locations))
+    : locationsForView;
+  const filteredLocations = search
+    ? regulatedLocations.filter((l) => SEARCH_FILTER(l, search))
+    : regulatedLocations;
+
+  const selectedGroups = groups
+    .filter(
+      (group) =>
+        getCheckedStateLocationGroup({
+          group,
+          selected: selectedLocations,
+          locations,
+        }) === "checked"
+    )
+    .map((g) => g.id);
+  const indeterminateGroups = groups
+    .filter(
+      (group) =>
+        getCheckedStateLocationGroup({
+          group,
+          selected: selectedLocations,
+          locations,
+        }) === "indeterminate"
+    )
+    .map((g) => g.id);
+
+  const selected = isGroupedView
+    ? [...selectedGroups, ...selectedLocations]
+    : selectedLocations;
+  const numSelected = selectedLocations.length;
 
   const handleChange = (newSelected: string[]) => {
     const updated = locations.map((location) => {
@@ -32,29 +83,73 @@ const LocationPickerCard = ({
     onChange(updated);
   };
 
-  return (
-    <PickerCard
-      title={title}
-      items={filteredLocations}
-      selected={selected}
-      onChange={handleChange}
-      toggle={
-        <Flex alignItems="center" gap="8px">
-          <Switch
-            isChecked={showRegulatedOnly}
-            size="sm"
-            onChange={() => setShowRegulatedOnly(!showRegulatedOnly)}
-            colorScheme="complimentary"
-            id={`${title}-regulated`}
-            data-testid="regulated-toggle"
-          />
-          <FormLabel fontSize="sm" m={0} htmlFor={`${title}-regulated`}>
-            Regulated
-          </FormLabel>
-          <QuestionTooltip label="Toggle on to see only locations in this region with privacy regulations supported by Fides" />
-        </Flex>
+  const handleGroupChange = (selections: string[]) => {
+    const newSelected = new Set(selections);
+    const oldSelected = new Set(selected);
+    // Handle additions
+    selections.forEach((s) => {
+      if (!oldSelected.has(s)) {
+        // If it's a group, we have to propagate
+        if (groups.find((g) => g.id === s)) {
+          locations
+            .filter((l) => l.belongs_to?.includes(s))
+            .forEach((l) => {
+              newSelected.add(l.id);
+            });
+        }
       }
-    />
+    });
+    // Handle removals
+    oldSelected.forEach((s) => {
+      if (!newSelected.has(s)) {
+        if (groups.find((g) => g.id === s)) {
+          locations
+            .filter((l) => l.belongs_to?.includes(s))
+            .forEach((l) => {
+              newSelected.delete(l.id);
+            });
+        }
+      }
+    });
+    handleChange(Array.from(newSelected));
+  };
+
+  if (filteredLocations.length === 0) {
+    return null;
+  }
+
+  return (
+    <>
+      <PickerCard
+        title={title}
+        items={filteredLocations}
+        selected={selected}
+        indeterminate={isGroupedView ? indeterminateGroups : []}
+        onChange={handleGroupChange}
+        onViewMore={() => {
+          disclosure.onOpen();
+        }}
+        numSelected={numSelected}
+        toggle={
+          <RegulatedToggle
+            id={`${title}-regulated`}
+            isChecked={showRegulatedOnly}
+            onChange={() => setShowRegulatedOnly(!showRegulatedOnly)}
+          />
+        }
+      />
+      <SubgroupModal
+        groups={groups}
+        locations={locations}
+        isOpen={disclosure.isOpen}
+        onClose={disclosure.onClose}
+        selected={selectedLocations}
+        onChange={handleChange}
+        // Rerender if a selection changes in this component so that the checkboxes
+        // in the modal stay up to date
+        key={`subgroup-modal-selected-${selected.length}`}
+      />
+    </>
   );
 };