+++ clients/admin-ui/src/pages/management/regulations.tsx
@@ -8,52 +8,60 @@ import {
   Text,
 } from "@fidesui/react";
 import type { NextPage } from "next";
-import Link from "next/link";
+import NextLink from "next/link";
 
 import { useAppSelector } from "~/app/hooks";
 import Layout from "~/features/common/Layout";
-import { USER_MANAGEMENT_ROUTE } from "~/features/common/nav/v2/routes";
-import LocationManagement from "~/features/locations/LocationManagement";
+import {
+  LOCATIONS_ROUTE,
+  USER_MANAGEMENT_ROUTE,
+} from "~/features/common/nav/v2/routes";
 import {
   selectLocationsRegulations,
   useGetLocationsRegulationsQuery,
 } from "~/features/locations/locations.slice";
+import RegulationManagement from "~/features/locations/RegulationManagement";
 
-const LocationsPage: NextPage = () => {
+const RegulationsPage: NextPage = () => {
   // Subscribe to locations/regulations endpoint
   const { isLoading } = useGetLocationsRegulationsQuery();
 
   const locationsRegulations = useAppSelector(selectLocationsRegulations);
 
   return (
-    <Layout title="Locations">
+    <Layout title="Regulations">
       <Box data-testid="location-management">
         <Heading marginBottom={2} fontSize="2xl">
-          Locations
+          Regulations
         </Heading>
         <Breadcrumb fontWeight="medium" fontSize="sm" mb="4">
           <BreadcrumbItem>
-            <Link href={USER_MANAGEMENT_ROUTE} passHref>
+            <NextLink href={USER_MANAGEMENT_ROUTE} passHref>
               <BreadcrumbLink href={USER_MANAGEMENT_ROUTE}>
                 Management
               </BreadcrumbLink>
-            </Link>
+            </NextLink>
           </BreadcrumbItem>
           <BreadcrumbItem>
-            <BreadcrumbLink href="#">Locations</BreadcrumbLink>
+            <BreadcrumbLink href="#">Regulations</BreadcrumbLink>
           </BreadcrumbItem>
         </Breadcrumb>
         <Box>
-          <Text marginBottom={4} fontSize="sm" maxWidth="720px">
-            Select the locations that you operate in and Fides will make sure
-            that you are automatically presented with the relevant regulatory
-            guidelines and global frameworks for your locations.
+          <Text marginBottom={4} fontSize="sm" maxWidth="600px">
+            Select the regulations that apply to your organizations compliance
+            requirements. The selections you make here will automatically update
+            your location selections.{" "}
+            <Text color="complimentary.500">
+              <NextLink href={LOCATIONS_ROUTE} passHref>
+                You can view your location settings here.
+              </NextLink>
+            </Text>
           </Text>
           <Box>
             {isLoading ? (
               <Spinner />
             ) : (
-              <LocationManagement data={locationsRegulations} />
+              <RegulationManagement data={locationsRegulations} />
             )}
           </Box>
         </Box>
@@ -61,4 +69,4 @@ const LocationsPage: NextPage = () => {
     </Layout>
   );
 };
-export default LocationsPage;
+export default RegulationsPage;